+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/commands/RemoveEntityCommand.java
@@ -10,6 +10,7 @@ import dev.qixils.crowdcontrol.socket.Response.ResultType;
 import lombok.Getter;
 import net.kyori.adventure.text.Component;
 import net.minecraft.core.registries.BuiltInRegistries;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.Entity.RemovalReason;
@@ -33,7 +34,7 @@ public class RemoveEntityCommand<E extends Entity> extends ImmediateCommand impl
 		super(plugin);
 		this.entityType = entityType;
 		this.effectName = "remove_entity_" + csIdOf(BuiltInRegistries.ENTITY_TYPE.getKey(entityType));
-		this.displayName = Component.translatable("cc.effect.remove_entity.name", entityType.getDescription());
+		this.displayName = Component.translatable("cc.effect.remove_entity.name", plugin.toAdventure(entityType.getDescription()));
 	}
 
 	@Override
@@ -44,8 +45,11 @@ public class RemoveEntityCommand<E extends Entity> extends ImmediateCommand impl
 	}
 
 	private boolean removeEntityFrom(ServerPlayer player) {
+		ServerLevel level = player.serverLevel();
+		if (entityType == EntityType.ENDER_DRAGON && level.getDragonFight() != null) return false;
+
 		Vec3 playerPosition = player.position();
-		List<Entity> entities = StreamSupport.stream(player.serverLevel().getAllEntities().spliterator(), false)
+		List<Entity> entities = StreamSupport.stream(level.getAllEntities().spliterator(), false)
 				.filter(entity -> entity.getType() == entityType && entity.distanceToSqr(playerPosition) <= REMOVE_ENTITY_RADIUS * REMOVE_ENTITY_RADIUS)
 				.sorted((entity1, entity2) -> (int) (entity1.distanceToSqr(playerPosition) - entity2.distanceToSqr(playerPosition))).toList();
 		if (entities.isEmpty())